@@ -9,29 +9,30 @@
   },
   "dependencies": {
     "@react-native-community/masked-view": "0.1.10",
-    "@react-native-firebase/app": "^12.4.0",
+    "@react-native-firebase/app": "^12.9.3",
     "@react-native-firebase/firestore": "^12.4.0",
     "@react-navigation/bottom-tabs": "^5.11.11",
     "@react-navigation/native": "^5.9.4",
     "@react-navigation/stack": "^5.14.5",
-    "expo": "^42.0.0",
-    "expo-image-picker": "~10.2.2",
-    "expo-status-bar": "~1.0.4",
-    "firebase": "8.2.3",
-    "react": "16.13.1",
-    "react-dom": "16.13.1",
-    "react-native": "https://github.com/expo/react-native/archive/sdk-42.0.0.tar.gz",
+    "expo": "^44.0.0",
+    "expo-image-picker": "~12.0.1",
+    "expo-status-bar": "~1.2.0",
+    "firebase": "^9.7.0",
+    "react": "17.0.1",
+    "react-dom": "17.0.1",
+    "react-native": "0.64.3",
     "react-native-camera-roll-picker": "^2.0.0",
     "react-native-elements": "^3.4.2",
-    "react-native-gesture-handler": "~1.10.2",
-    "react-native-reanimated": "~2.2.0",
-    "react-native-safe-area-context": "3.2.0",
-    "react-native-screens": "~3.4.0",
-    "react-native-web": "~0.13.12",
+    "react-native-gesture-handler": "~2.1.0",
+    "react-native-image-picker": "^4.8.2",
+    "react-native-reanimated": "~2.3.1",
+    "react-native-safe-area-context": "3.3.2",
+    "react-native-screens": "~3.10.1",
+    "react-native-web": "0.17.1",
     "react-navigation": "^4.4.4"
   },
   "devDependencies": {
-    "@babel/core": "~7.9.0"
+    "@babel/core": "^7.12.9"
   },
   "private": true
 }
