@@ -6,8 +6,6 @@ import { Button, Image, Input } from "react-native-elements"
 import { NavigationContainer } from '@react-navigation/native';
 import { createStackNavigator } from '@react-navigation/stack';
 import { StackActions } from '@react-navigation/native';
-import * as firebase from 'firebase'
-import { auth } from '../firebase'
 import { ScrollView } from 'react-native-gesture-handler'
 import ProfileRecentPosts from '../components/ProfileRecentPosts'
 import { TouchableOpacity } from 'react-native'
@@ -15,6 +13,7 @@ import { useState } from 'react'
 import ChangeName from '../components/ChangeName'
 import "firebase/firestore";
 import { db } from '../firebase'
+import { auth } from '../firebase'
 import UploadPhoto from '../components/UploadPhoto'
 
 const AccountSettings = ({ navigation }) => {
@@ -22,7 +21,7 @@ const AccountSettings = ({ navigation }) => {
     const [isShown, setIsShown] = useState(false)
     const [uploadAvatar, setUploadAvatar] = useState(false)
     const [changedName, setChangedName] = useState("") 
-    const user = firebase.auth().currentUser;
+    const user = auth.currentUser;
 
     const changeName = () => {
         setIsShown(true)
